+++ src-dbg/flow_storm/debugger/events_processor.clj
@@ -3,12 +3,9 @@
   "Processing events the debugger receives from the runtime"
 
   (:require [flow-storm.debugger.ui.browser.screen :as browser-screen]
-            [flow-storm.debugger.ui.flows.general :as ui-general]
             [flow-storm.debugger.ui.taps.screen :as taps-screen]
             [flow-storm.debugger.ui.main :as ui-main]
             [flow-storm.debugger.ui.flows.screen :as flows-screen]
-            [flow-storm.debugger.runtime-api :as runtime-api :refer [rt-api]]
-            [flow-storm.debugger.ui.flows.code :as ui-code]
             [flow-storm.debugger.ui.docs.screen :as docs-screen]
             [flow-storm.debugger.config :refer [debug-mode]]
             [flow-storm.debugger.ui.utils :as ui-utils]
@@ -18,27 +15,27 @@
 (defn- var-instrumented-event [{:keys [var-ns var-name]}]
   (ui-utils/run-later
    (browser-screen/add-to-instrumentation-list (browser-screen/make-inst-var var-ns var-name))
-   (ui-main/select-main-tools-tab :browser)))
+   (ui-vars/select-main-tools-tab :browser)))
 
 (defn- var-uninstrumented-event [{:keys [var-ns var-name]}]
   (ui-utils/run-later
    (browser-screen/remove-from-instrumentation-list (browser-screen/make-inst-var var-ns var-name))
-   (ui-main/select-main-tools-tab :browser)))
+   (ui-vars/select-main-tools-tab :browser)))
 
 (defn- namespace-instrumented-event [{:keys [ns-name]}]
   (ui-utils/run-later
    (browser-screen/add-to-instrumentation-list (browser-screen/make-inst-ns ns-name))
-   (ui-main/select-main-tools-tab :browser)))
+   (ui-vars/select-main-tools-tab :browser)))
 
 (defn- namespace-uninstrumented-event [{:keys [ns-name]}]
   (ui-utils/run-later
    (browser-screen/remove-from-instrumentation-list (browser-screen/make-inst-ns ns-name))
-   (ui-main/select-main-tools-tab :browser)))
+   (ui-vars/select-main-tools-tab :browser)))
 
 (defn- tap-event [{:keys [value]}]
   (ui-utils/run-later
    (taps-screen/add-tap-value value)
-   (ui-main/select-main-tools-tab :taps)))
+   (ui-vars/select-main-tools-tab :taps)))
 
 (defn- flow-created-event [flow-info]
   (ui-utils/run-now
@@ -61,21 +58,13 @@
 (defn- heap-info-update-event [ev-args-map]
   (ui-main/update-heap-indicator ev-args-map))
 
-(defn- goto-location-event [{:keys [flow-id thread-id thread-name idx]}]
+(defn- goto-location-event [loc]
   (ui-utils/run-now
-   (ui-main/select-main-tools-tab :flows)
-   (flows-screen/select-flow-tab flow-id)
-   (flows-screen/open-thread {:flow/id flow-id
-                              :thread/id thread-id
-                              :thread/name thread-name})
-   (ui-general/select-thread-tool-tab flow-id thread-id :code)
-   (ui-code/jump-to-coord flow-id
-                          thread-id
-                          (runtime-api/timeline-entry rt-api flow-id thread-id idx :at))))
+   (flows-screen/goto-location loc)))
 
 (defn- show-doc-event [{:keys [var-symbol]}]
   (ui-utils/run-now
-   (ui-main/select-main-tools-tab :docs)
+   (ui-vars/select-main-tools-tab :docs)
    (docs-screen/show-doc var-symbol)))
 
 (defn- break-installed-event [{:keys [fq-fn-symb]}]