+++ src-dbg/flow_storm/debugger/ui/taps/screen.clj
@@ -3,6 +3,7 @@
              :as ui-utils
              :refer [button label list-view h-box v-box]]
             [flow-storm.debugger.ui.state-vars :refer [store-obj obj-lookup] :as ui-vars]
+            [flow-storm.debugger.ui.flows.screen :as flows-screen]
             [flow-storm.debugger.runtime-api :as runtime-api :refer [rt-api]]
             [flow-storm.debugger.ui.value-inspector :as value-inspector])
   (:import [javafx.scene.layout Priority VBox]
@@ -19,6 +20,12 @@
    (let [[{:keys [clear]}] (obj-lookup "taps-list-view-data")]
      (clear))))
 
+(defn find-and-jump-tap-val [vref]
+  (when-let [tentry (runtime-api/find-timeline-entry rt-api {:from-idx 0
+                                                             :eq-val-ref vref})]
+    (ui-vars/select-main-tools-tab :flows)
+    (flows-screen/goto-location tentry)))
+
 (defn main-pane []
   (let [{:keys [list-view-pane] :as lv-data}
         (list-view {:editable? false
@@ -29,17 +36,27 @@
                                                                                                          :pprint? false}))]
                                          (.setText list-cell nil)
                                          (.setGraphic list-cell (label val-list-text))))
-                    :on-click (fn [mev sel-items _]
-                                (when (and (= MouseButton/PRIMARY (.getButton mev))
-                                           (= 2 (.getClickCount mev)))
-                                  (let [val (first sel-items)]
-                                    (value-inspector/create-inspector val {}))))
+                    :on-click (fn [mev sel-items {:keys [list-view]}]
+                                (let [val (first sel-items)]
+                                  (cond
+                                    (and (= MouseButton/PRIMARY (.getButton mev))
+                                         (= 2 (.getClickCount mev)))
+                                    (value-inspector/create-inspector val {})
+
+                                    (= MouseButton/SECONDARY (.getButton mev))
+                                    (let [ctx-menu (ui-utils/make-context-menu [{:text "Search value on Flows"
+                                                                                 :on-click (fn [] (find-and-jump-tap-val val))}])]
+                                      (.show ctx-menu
+                                             list-view
+                                             (.getScreenX mev)
+                                             (.getScreenY mev))))))
                     :selection-mode :single})
         clear-btn (button :label "clear"
                           :on-click (fn [] (clear-all-taps)))
         header-pane (doto (h-box [clear-btn])
                       (.setAlignment Pos/CENTER_RIGHT))
         mp (v-box [header-pane list-view-pane])]
+
 
     (VBox/setVgrow list-view-pane Priority/ALWAYS)
     (VBox/setVgrow mp Priority/ALWAYS)